refactor(theme): extract app-shell colors into a ui-theme util lib; drop util/
The app-shell COLOR palette lived in a loose util/theme.js reached via the
util/* alias. Move it into a proper Nx util lib so the design tokens have a
tagged, boundary-checked home the app (and any future lib) can import.

- new libs/ui-theme (tag type:util) — the first util lib; exports COLOR from
  @my-portfolio/ui-theme
- repoint the 7 components/ imports 'util/theme' -> '@my-portfolio/ui-theme'
- alias @my-portfolio/ui-theme in tsconfig.json + tsconfig.base.json + jest
- delete util/ entirely: theme.js moved out; progress.js was dead (only a
  commented-out import in _app.tsx) so drop it, the nprogress dependency, and
  the commented Router.onRouteChange blocks
- remove the now-unused util/* alias from tsconfig + jest + boundary allow-list
- update CLAUDE.md

Each mini-app keeps its own brand palette; ui-theme is deliberately just the
site-shell tokens, not a merged cross-app theme.

Verified: nx typecheck 13/13, nx lint 0 errors (app->util boundary OK), root
eslint 0 errors, build 24/24, testc 3 suites / 7 tests.
