+++ libs/ui-theme/src/index.ts
@@ -1,0 +1,2 @@
+// Public API of the ui-theme lib: the portfolio site's shared design tokens.
+export { COLOR } from './lib/colors';
