+++ package.json
@@ -18,7 +18,6 @@
     "gsap": "3.15.0",
     "next": "15",
     "next-redux-wrapper": "8.1.0",
-    "nprogress": "^0.2.0",
     "prop-types": "15.8.1",
     "react": "19",
     "react-dom": "19",
