+++ tsconfig.json
@@ -30,8 +30,8 @@
       "@my-portfolio/haptik": ["libs/haptik/src/index.ts"],
       "@my-portfolio/plaza": ["libs/plaza/src/index.ts"],
       "@my-portfolio/appbase": ["libs/appbase/src/index.ts"],
+      "@my-portfolio/ui-theme": ["libs/ui-theme/src/index.ts"],
       "components/*": ["components/*"],
-      "util/*": ["util/*"],
       "store": ["store"],
       "store/*": ["store/*"],
       "images/*": ["public/images/*"]
