--- util/theme.js
@@ -1,11 +1,0 @@
-export const COLOR = {
-  BLUE: '#337AB7',
-  ORANGE: '#F97168',
-  LIGHT_GREY: '#F2F2F2',
-  YELLOW: '#FFC600',
-  LIGHT_YELLOW: '#FFF3D8',
-  WHITE: '#FFFFFF',
-  BLACK: '#000000',
-};
-
-export const BOX_SHADOW = {};
